+++ compute-worker/src/api/operation-keys.ts
@@ -1,5 +1,5 @@
 import { createHash } from 'node:crypto';
-import { PDF_PARSER_VERSION } from '../inference/pdf/parser-version';
+import { PDF_PARSER_VERSION } from '../api/contracts';
 
 function sha256Hex(input: string): string {
   return createHash('sha256').update(input).digest('hex');
