--- compute-worker/src/inference/index.ts
@@ -1,26 +1,0 @@
-export * from '../api/contracts';
-export {
-  getComputeJobConcurrency,
-  getAvailableCpuCores,
-  getOnnxThreadsPerJob,
-} from './config/cpu-budget';
-export {
-  getComputeTimeoutConfig,
-  getComputeOpStaleMs,
-  getWorkerClientWaitTimeoutMs,
-  withTimeout,
-  withIdleTimeoutAndHardCap,
-  type ComputeTimeoutConfig,
-  type ComputeOperationKind,
-  type IdleTimeoutAndHardCapInput,
-} from './config/timeout';
-export { renderPage } from './pdf/render';
-export { mergeTextWithRegions } from './pdf/merge';
-export { PDF_PARSER_VERSION } from './pdf/parser-version';
-export { encodeParserVersion } from './pdf/parser-version-key';
-export { stitchCrossPageBlocks } from './pdf/stitch';
-export { normalizeTextItemsForLayout } from './pdf/normalize-text';
-export { mapWordsToSentenceOffsets, type WhisperWord } from './whisper/alignment-map';
-export { buildGoertzelCoefficients, goertzelPower } from './whisper/spectral';
-export { buildWordsFromTimestampedTokens, extractTokenStartTimestamps } from './whisper/token-timestamps';
-export * from '../operations';
