--- compute-worker/src/inference/pdf/parser-version-key.ts
@@ -1,9 +1,0 @@
-import { PDF_PARSER_VERSION } from './parser-version';
-
-export function encodeParserVersion(
-  parserVersion: string,
-  defaultVersion = PDF_PARSER_VERSION,
-): string {
-  const normalized = parserVersion.trim() || defaultVersion;
-  return encodeURIComponent(normalized);
-}
