--- compute-worker/src/inference/pdf/parser-version.ts
@@ -1,1 +1,0 @@
-export const PDF_PARSER_VERSION = 'pp-doclayoutv3-onnx@800+pdfjs@4.8.69';
