+++ compute-worker/src/inference/pdf/pdfjs.ts
