--- compute-worker/src/inference/pdf/types.ts
@@ -1,15 +1,0 @@
-import type { ParsedPdfBlockKind } from '../types/parsed-pdf';
-
-export interface PdfTextItem {
-  text: string;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-export interface LayoutRegion {
-  bbox: [number, number, number, number];
-  label: ParsedPdfBlockKind;
-  confidence?: number;
-}
