+++ compute-worker/src/inference/runtime.ts
