--- compute-worker/src/inference/types/index.ts
@@ -1,17 +1,0 @@
-export type {
-  TTSAudioBuffer,
-  TTSAudioBytes,
-  TTSSentenceAlignment,
-  TTSSentenceWord,
-} from './tts';
-
-export type {
-  ParsedPdfBlock,
-  ParsedPdfBlockFragment,
-  ParsedPdfBlockKind,
-  ParsedPdfDocument,
-  ParsedPdfPage,
-  PdfParsePhase,
-  PdfParseProgress,
-  PdfParseStatus,
-} from './parsed-pdf';
