--- compute-worker/src/inference/whisper/alignment-map.ts
@@ -1,44 +1,0 @@
-import type { TTSSentenceAlignment, TTSSentenceWord } from '../types/tts';
-
-export interface WhisperWord {
-  start: number;
-  end: number;
-  word: string;
-}
-
-export function mapWordsToSentenceOffsets(sentence: string, words: WhisperWord[]): TTSSentenceAlignment {
-  const lowerSentence = sentence.toLowerCase();
-  let cursor = 0;
-
-  const alignedWords: TTSSentenceWord[] = words.map((w) => {
-    const token = w.word.trim();
-    if (!token) {
-      return {
-        text: '',
-        startSec: w.start,
-        endSec: w.end,
-        charStart: cursor,
-        charEnd: cursor,
-      };
-    }
-
-    const idx = lowerSentence.indexOf(token.toLowerCase(), cursor);
-    const start = idx >= 0 ? idx : cursor;
-    const end = Math.min(sentence.length, start + token.length);
-    cursor = Math.max(cursor, end);
-
-    return {
-      text: token,
-      startSec: w.start,
-      endSec: w.end,
-      charStart: start,
-      charEnd: end,
-    };
-  }).filter((word) => word.text.length > 0);
-
-  return {
-    sentence,
-    sentenceIndex: 0,
-    words: alignedWords,
-  };
-}
