+++ compute-worker/src/infrastructure/platform.ts
@@ -1,5 +1,25 @@
-import { existsSync } from 'fs';
+import fs from 'fs';
+import path from 'path';
 import ffmpegStatic from 'ffmpeg-static';
+
+function findMonorepoRoot(startDir: string): string | null {
+  let current = path.resolve(startDir);
+  for (;;) {
+    const marker = path.join(current, 'pnpm-workspace.yaml');
+    if (fs.existsSync(marker)) return current;
+    const parent = path.dirname(current);
+    if (parent === current) return null;
+    current = parent;
+  }
+}
+
+function resolveDocstoreDir(): string {
+  const repoRoot = findMonorepoRoot(process.cwd());
+  if (repoRoot) return path.join(repoRoot, 'docstore');
+  return path.join(process.cwd(), 'docstore');
+}
+
+export const DOCSTORE_DIR = resolveDocstoreDir();
 
 function normalizePath(value: unknown): string | null {
   if (typeof value !== 'string') return null;
@@ -9,20 +29,17 @@ function normalizePath(value: unknown): string | null {
 
 function resolveBinary(envValue: string | null, bundledValue: string | null, envVarName: string, packageName: string): string {
   if (envValue) {
-    if ((envValue.includes('/') || envValue.includes('\\')) && !existsSync(envValue)) {
+    if ((envValue.includes('/') || envValue.includes('\\')) && !fs.existsSync(envValue)) {
       throw new Error(`${envVarName} points to a missing binary: ${envValue}`);
     }
     return envValue;
   }
-
   if (!bundledValue) {
     throw new Error(`${packageName} binary is unavailable on this platform. Set ${envVarName} to an installed binary path.`);
   }
-
-  if ((bundledValue.includes('/') || bundledValue.includes('\\')) && !existsSync(bundledValue)) {
+  if ((bundledValue.includes('/') || bundledValue.includes('\\')) && !fs.existsSync(bundledValue)) {
     throw new Error(`${packageName} resolved to a missing binary path: ${bundledValue}`);
   }
-
   return bundledValue;
 }
